pyproject.toml
[project]
name = "mamba-ettm1-forecasting"
version = "0.1.0"
description = "Mamba-based long-horizon time series forecasting on ETTm1 dataset"
authors = [
    { name = "Stein Pleiter", email = "stein.pleiter@mail.utoronto.ca" }
]
dependencies = [
    "torch>=2.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]
readme = "README.md"
requires-python = ">= 3.9"
license = { text = "MIT" }
keywords = ["time-series", "forecasting", "mamba", "deep-learning", "state-space-models"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
# Mamba dependencies (Linux/CUDA only - not available on macOS)
mamba = [
    "mamba-ssm>=1.2.0",
    "causal-conv1d>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/steinpleiter/Mamba-ETTm1-Forecasting"
Repository = "https://github.com/steinpleiter/Mamba-ETTm1-Forecasting"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    # Development tools
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    
    # Notebooks
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
    "ipykernel>=6.25.0",
]

[tool.rye.scripts]
preprocess = "python scripts/preprocess_data.py"
train = "python scripts/train.py"
evaluate = "python scripts/evaluate.py"
test = "pytest tests/"
format = { chain = ["format:black", "format:isort"] }
"format:black" = "black src/ tests/ scripts/"
"format:isort" = "isort src/ tests/ scripts/"
lint = { chain = ["lint:black", "lint:flake8"] }
"lint:black" = "black --check src/ tests/ scripts/"
"lint:flake8" = "flake8 src/ tests/ scripts/"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
